+++ codegen.c
@@ -194,7 +194,7 @@ void gen_global_init(VariableInit *init, Type *type) {
     if (init->expr) {
       if (type->base->kind == TYPE_CHAR && init->expr->kind == ND_VAR && init->expr->variable->kind == VK_STRLIT) {
         char *lit = init->expr->variable->string_literal;
-        if (type->array_size != strlen(lit) + 1)
+        if (type->array_size != (int)strlen(lit) + 1)
           error_at(init->expr->token->pos, "miss-match between array-size and string-length");
         printf("  .ascii \"%s\\0\"\n", lit);
         return;
